WS1: extract EmojiOverlay into its own component module
Fourth slice of the modelmap.tsx split (1,000 -> 913). Move the EmojiOverlay
canvas overlay (hotspot-pulse + emoji rendering driven by the map render loop)
into src/features/model-map/emoji-overlay.tsx as a default-exported client
component. applyAlpha was only used here, so it drops from the component's
map-constants import.

Verbatim move, behavior-preserving: tsc clean, lint clean, pnpm test 23/23.
